ecotaxa/ecotaxa_dev#693: Add an optional path to uploaded files. Try to boost file transfer by caching existing dirs and avoiding rewrite.

+++ py/FS/UserDir.py
@@ -5,34 +5,64 @@
 import os.path
 import tempfile
 from pathlib import Path
-from typing import Any
+from typing import Any, Optional, List, Set
 
 from BO.User import UserIDT
+from FS.CommonDir import CommonFolder, DirEntryT
+from FS.TempDirForTasks import TempDirForTasks
 
 
 class UserDirectory(object):
     """
         Base directory for storing user files. So far just a temp space.
     """
+    USER_DIR_PATTERN = "ecotaxa_user.%d"
+    user_dir_cache: Set[Path] = set()
 
-    def __init__(self, user_id: UserIDT):
+    def __init__(self, user_id: UserIDT, tag: Optional[str] = None):
         self.user_id = user_id
+        self.tag = tag
 
     def _user_suffix(self):
         return "." + str(self.user_id)
 
-    async def add_file(self, name: str, stream: Any) -> str:
+    async def add_file(self, name: str, path: Optional[str], stream: Any) -> str:
         """
             Add the byte stream as the file with name 'name' into self.
+            :param name: File name.
+            :param path: The client-side full path of the file. For replicating a directory structure.
+            :param stream: The byte stream with file content.
         """
-        path: Path = Path(tempfile.mkdtemp(suffix=self._user_suffix()))
-        dest_path = path.absolute().joinpath(name)
+        base_path: Path
+        if self.tag is not None:
+            base_path = Path(tempfile.gettempdir(), self.USER_DIR_PATTERN % self.user_id)
+            TempDirForTasks.ensure_exists(base_path, self.user_dir_cache)
+            base_path /= self.tag
+        else:
+            base_path = Path(tempfile.mkdtemp(suffix=self._user_suffix()))
+        if path is not None:
+            assert path.endswith(name)
+            base_path /= path[:-len(name)]
+        TempDirForTasks.ensure_exists(base_path, self.user_dir_cache)
+        dest_path = base_path.absolute().joinpath(name)
+        # TODO: Should be an option
+        if dest_path.exists():
+            return str(dest_path)
         with open(dest_path, "wb") as fout:
             buff = await stream.read(1024)
             while len(buff) != 0:
                 fout.write(buff)
                 buff = await stream.read(1024)
         return str(dest_path)
+
+    def list(self, sub_path: str) -> List[DirEntryT]:
+        """
+            Only list the known (with tags) directory.
+        """
+        ret: List[DirEntryT] = []
+        path: Path = Path(tempfile.gettempdir(), self.USER_DIR_PATTERN % self.user_id, sub_path)
+        CommonFolder.list_dir_into(path, ret)
+        return ret
 
     def contains(self, path_str: str) -> bool:
         """
